create_model.py is not generated by the 'main.rs'. It is invoked by it
